pyproject.toml
[tool.poetry]
name = "gfmodules-revocation-service"
version = "0.1.0"
description = ""
license = "EUPL-1.2"
authors = ["Ministerie van Volksgezondheid, Welzijn en Sport"]
repository = "https://github.com/minvws/gfmodules-generieke-issuer-revocatie-api-private"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.35.0"
fastapi = "^0.115.12"
mypy = "^1.16.0"
inject = "^5.2.0"
opentelemetry-sdk = "^1.34.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.34.0"
opentelemetry-instrumentation = "^0.55b1"
opentelemetry-instrumentation-fastapi = "^0.55b1"
opentelemetry-instrumentation-requests = "^0.55b1"
requests = "^2.32.0"
statsd = "^4.0.1"
poetry-core = "^2.1.3"
pydantic-settings = "^2.9.1"
redis = "^6.2.0"
urllib3 = "^2.5.0"
sqlalchemy = "^2.0.41"
psycopg = "^3.2.9"
pyjwt = "^2.10.1"
cryptography = "^45.0.4"
serialization = "^1.2.3"
base58 = "^2.1.1"
multiformats = "^0.3.1.post4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.1"
httpx = "^0.28.1"
ruff = "0.11.12"
safety = "^3.2.14"
codespell = "^2.4.1"
types-requests = "^2.32.4.20250611"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
cache-dir = "~/.cache/ruff"
include = ["pyproject.toml", "app/*.py", "tests/*.py"]
line-length = 120

[tool.mypy]
files = "app,tests"
python_version = "3.11"
strict = true
cache_dir = "~/.cache/mypy"
ignore_missing_imports = true

[tool.pytest.ini_options]
cache_dir = "~/.cache/pytest"
testpaths = ["tests"]
